BugFix : 유저 리워드 적립 차감 로직 수정

+++ src/main/java/com/sparta/hh99_actualproject/model/Member.java
@@ -45,7 +45,7 @@ public class Member {
     private String kakaoUserId;
 
     @Column
-    private Integer reward;
+    private Float reward;
 
     @OneToMany(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Message> messageList;
